requirements.txt
aiohttp==3.9.3
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
arrow==1.2.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
binaryornot==0.4.4
CacheControl==0.14.0
cachetools==5.3.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.5
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-appconf==1.0.5
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.1
django-compressor==4.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-froala-editor==4.0.15
django-grappelli==3.0.7
django-htmx==1.16.0
django-import-export==3.0.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-libsass==0.9
django-nextjs==2.2.1
django-passkeys==1.2.7
django-resized==1.0.2
django-rest-swagger==2.2.0
django-sass-processor==1.2.1
django-simpleui==2023.8.28
django-solo==2.0.0
django-tailwind==3.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.1.1
fido2==1.1.1
firebase-admin==6.4.0
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
google-trans-new==1.1.9
googleapis-common-protos==1.62.0
googletrans==3.0.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inertia-django==0.5.2
inflection==0.5.1
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.2.4
libsass==0.21.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
midtransclient==1.4.1
mnemonic==0.20
msgpack==1.0.7
multidict==6.0.2
mysql-client==0.0.1
mysqlclient==2.2.0
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.0
python3-openid==3.2.0
pytube==12.1.0
pytz==2022.5
PyYAML==6.0
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rjsmin==1.2.0
RouterOS-api==0.17.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.5.1
stellar-base-sseclient==0.0.21
stellar-sdk==8.2.1
tablib==3.2.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.5.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2022.5
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
websockets==10.4
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.1
django-loginas
